SGBUSAND-180 Moves inner AsyncTasks to its own class
This will prevent static field leaks and other issues happening as well as segregate the async tasks out properly

Resolves SGBUSAND-180
